ANDROID: fs-verity: Fix up function signatures to match ACK
In cherry-picking "ANDROID: fs-verity: Export function to check signatures" to
pixel-4.19 branches, the signature of __fsverity_verify_signature was
changed. Change back to original signature to avoid GKI breakage

Bug: 186855533
Test: incfs_test passes, GKI builds match

+++ fs/verity/signature.c
@@ -61,7 +61,7 @@ int fsverity_verify_signature(const struct fsverity_info *vi,
  * Return: 0 on success (signature valid or not required); -errno on failure
  */
 int __fsverity_verify_signature(const struct inode *inode, const u8 *signature,
-				size_t sig_size, const u8 *file_digest,
+				u32 sig_size, const u8 *file_digest,
 				unsigned int digest_algorithm)
 {
 	struct fsverity_formatted_digest *d;
